feat: prompt video filter
